% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_synthetic_data.R
\name{generate_synthetic_data}
\alias{generate_synthetic_data}
\title{Generate Synthetic Data}
\usage{
generate_synthetic_data(n_samples, seed = NULL)
}
\arguments{
\item{n_samples}{Integer. Number of samples to generate.}
}
\value{
A data frame containing the synthetic dataset.
}
\description{
This function generates synthetic data using parameters sourced from the \code{exported_parameters} object.
#'
}
\examples{
synthetic_data <- generate_synthetic_data(n_samples = 1000)
}
